fix: update SettingDialog to use Modal component and stop using styles from calcite-web.js
