--- src/components/SettingDialog/style.css
@@ -1,3 +1,0 @@
-.setting-dialog .btn-transparent {
-    color: #fff;
-}
